--- src/pages/main-empty-layout.tsx
@@ -1,11 +1,0 @@
-import { Outlet } from "react-router-dom"
-
-function MainEmptyLayout() {
-  return (
-    <div>
-      <Outlet />
-    </div>
-  )
-}
-
-export default MainEmptyLayout
